+++ app/containers/index/travel-stats/travel-stats.tsx
@@ -1,17 +1,13 @@
-import { /* useState, */ type FC } from 'react';
-// import clsx from 'classnames';
+import { type FC } from 'react';
 import {
     TRAVELLED_COUNTRY_COUNT,
     TRAVELLED_COUNTRY_COUNT_BY_REGION,
     type CountryInfo,
     type WorldRegion,
 } from '~/data-access/country';
-// import { Button } from '~/components/button';
-// import { ChevronDoubleDownOutlineIcon } from '~/icons/outline/chevron-double-down';
-// import { ChevronDoubleUpOutlineIcon } from '~/icons/outline/chevron-double-up';
 import { trim } from '~/utils/trim';
 import { TravelledCountriesCounter } from './travelled-countries-counter';
-import { CountryFlagList } from './travelled-countries-flag-list.desktop';
+import { CountryFlagList } from './travelled-countries-flag-list';
 import { RegionBackgroundWrapper } from './region-background-wrapper';
 import './travel-stats.css';
 
@@ -76,73 +72,13 @@ export const TravelStats: FC<TravelStatsProps> = ({ className }) => {
                         </>
                     }
                 />
-                {/**
-                 *  We show the country flags with tooltip on desktop version
-                 *  only since there's enough space
-                 */}
-
                 <CountryFlagList
+                    className="travel-stats__country-flag-list"
                     countries={
                         TRAVELLED_COUNTRY_COUNT_BY_REGION['Europe'].countries
                     }
                 />
             </RegionBackgroundWrapper>
-
-            {/* TODO: Redesign Mobile View */}
-            {/**
-             *  For mobile view, since it is impossible to fit country flag list with the
-             *  visited country banner section, the following populates a collapsable list
-             *  of visited country info
-             */}
-            {/* <div className='px-4 pt-8 text-center bg-blue-500 text-white sm:hidden'>
-                <ul className={clsx(
-                    'travel-stats__country-list',
-                    { 'travel-stats__country-list--expanded': expandedSections.has('Europe') }
-                )}>
-                    {TRAVELLED_COUNTRY_COUNT_BY_REGION['Europe'].countries.map(({
-                        countryCode,
-                        fullname,
-                        name
-                    }) => (
-                        <li key={countryCode}>
-                            <span className='inline-block'>
-                            <CountryFlagIcon
-                                countryCode={countryCode}
-                                size='md'
-                                className='rounded'
-                            />
-                            </span>
-
-                            <p className='font-header font-bold text-sm'>
-                                {name}
-                            </p>
-                            <p className='text-xs text-white/70'>
-                                {fullname}
-                            </p>
-                        </li>
-                    ))}
-                </ul>
-
-                <Button
-                    size='sm'
-                    variant='tertiary'
-                    startIcon={expandedSections.has('Europe') ? ChevronDoubleUpOutlineIcon : ChevronDoubleDownOutlineIcon}
-                    endIcon={expandedSections.has('Europe') ? ChevronDoubleUpOutlineIcon : ChevronDoubleDownOutlineIcon}
-                    className='!text-white w-[calc(100%+2rem)] justify-center ml-[-1rem] !bg-blue-400/40'
-                    onClick={() => {
-                        const updated = new Set(expandedSections);
-                        if (updated.has('Europe')) {
-                            updated.delete('Europe');
-                        } else {
-                            updated.add('Europe');
-                        }
-
-                        setExpandedSections(updated);
-                    }}
-                >
-                    {expandedSections.has('Europe') ? 'COLLAPSE' : 'EXPAND MORE'}
-                </Button>
-            </div> */}
 
             <RegionBackgroundWrapper
                 className="travel-stats__bg travel-stats__asia-bg"
